RecetasChidas/requirements.txt
Django==4.2.20
python-decouple==3.8
mysqlclient==2.1.1
